// apps/web/lib/generated/prisma/models/ShiftType.ts
export type ShiftTypeModel = runtime.Types.Result.DefaultSelection<Prisma.$ShiftTypePayload>
// ...
export type $ShiftTypePayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "ShiftType"
  objects: {
    team: Prisma.$TeamPayload<ExtArgs>
    assignments: Prisma.$AssignmentPayload<ExtArgs>[]
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    name: string
    startTime: string
    endTime: string
    color: string
    teamId: string
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["shiftType"]>
  composites: {}
}